+++ src/components/organisms/Header.jsx
@@ -1,45 +1,57 @@
-import React from 'react';
+import { useEffect, useContext } from 'react';
 import styled from '@emotion/styled';
 
-const HeaderWrapper = styled.div`
-  position: fixed;
-  display: flex;
-  justify-content: space-between;
-  color: white;
-  font-family: 'poppins', sans-serif;
-  border: 1px solid #214eea;
-  background: #214eea;
-  padding: 1rem 2rem;
-  align-items: center;
-  position: fixed;
-  top: 0;
-  z-index: 2;
-  width: 100%;
-`;
+import GlobalContext from 'context/GlobalContext';
+import NavMenu from 'molecules/NavMenu';
+import MobileNavMenu from 'organisms/MobileNavMenu';
+import MobileMenuButton from 'atoms/MobileMenuButton';
+import links from 'data/navigation';
 
-const StyledLogo = styled.div`
-  font-weight: Bold;
-  font-size: 30px;
-`;
+const SampleHeader = () => {
+  const { scrolled, handleScroll } = useContext(GlobalContext);
+  const { menuOpen, toggleMenu } = useContext(GlobalContext);
 
-const StyledNav = styled.div`
-  display: none;
-  @media (min-width: 768px) {
-    display: flex;
-    justify-content: center;
-    align-items: flex-end;
-    font-size: 12px;
-    margin: 0;
-  }
-`;
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, []);
 
-const Header = () => {
   return (
-    <HeaderWrapper>
-      <StyledLogo>Transpiled</StyledLogo>
-      <StyledNav>Services</StyledNav>
-    </HeaderWrapper>
+    <HeaderContainer scrolled={scrolled}>
+      <MobileNavMenu links={links} />
+      <Logo>Transpiled</Logo>
+      <Nav>
+        <NavMenu links={links} />
+        <MobileMenuButton onClick={toggleMenu} isOpen={menuOpen} />
+      </Nav>
+    </HeaderContainer>
   );
 };
 
-export default Header;
+const HeaderContainer = styled.header`
+  position: fixed;
+  top: 0;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 1rem 2rem;
+  background-color: ${({ scrolled }) => (scrolled ? 'black' : 'transparent')};
+  color: ${({ theme }) => theme.colors.white};
+  transition:
+    background-color 0.3s,
+    color 0.3s;
+  z-index: ${({ theme }) => theme.zIndices.header};
+`;
+
+const Logo = styled.div`
+  font-size: 2rem;
+  font-weight: bold;
+`;
+
+const Nav = styled.nav`
+  display: flex;
+  align-items: center;
+`;
+
+export default SampleHeader;
