+++ src/components/templates/PageLayout.jsx
@@ -1,10 +1,9 @@
 // components/Layout.js
 import PropTypes from 'prop-types';
-import Header from 'organisms/SampleHeader';
+import Header from 'organisms/Header';
 import Footer from 'organisms/Footer';
 
 const Layout = ({ children }) => {
-
   return (
     <>
       <Header />
