+++ src/data/navigation.js
@@ -1,7 +1,7 @@
 const links = [
   { url: '/', label: 'Home' },
   { url: '/services', label: 'Services' },
-  { url: '/testimonials', label: 'Testimonials' },
+  // { url: '/testimonials', label: 'Testimonials' },
   { url: '/contact', label: 'Contact' },
   { url: '/about', label: 'About' },
 ];
